LayoutTests/storage/indexeddb/modern/objectstore-autoincrement-types-expected.txt

Test IndexedDB's IDBObjectStore auto-increment feature with more types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
createObjectStore():
store = db.createObjectStore('Store', {keyPath: 'primaryKey', autoIncrement: true})
PASS request.result.size is blob.size
PASS request.result.type is blob.type
PASS request.result.primaryKey is primaryKey
PASS request.result.name is file.name
PASS request.result.lastModified is file.lastModified
PASS request.result.primaryKey is primaryKey
PASS request.result.height is imageData.height
PASS request.result.width is imageData.width
PASS JSON.stringify(request.result.data) is JSON.stringify(imageData.data)
PASS request.result.primaryKey is primaryKey
PASS request.result.length is fileList.length
PASS request.result.primaryKey is primaryKey
PASS successfullyParsed is true

TEST COMPLETE
